Fast fail + better error reporting
The build bots:

- Used to run all tests, even if the first one failed.
- Only display the last 20 log lines during execution,
  which used to only contain a stacktrace indicating failure, not which
  test actually failed.

We now extract a test failure if we have one, and format it so we can
see it immediately in the failure snippet.

+++ pytest/test_embedded/tests/conftest.py
@@ -98,6 +98,11 @@ def pytest_addoption(parser):
         default=30,
         help="Number of seconds the frame perf test should last.",
     )
+    parser.addoption(
+        "--avd_keep",
+        action="store_true",
+        help="Do not delete the created avds. Useful if you need to debug snapshot related issues.",
+    )
 
 
 ALL_PLATFORMS = set("darwin linux win32".split())
@@ -161,7 +166,8 @@ def pytest_sessionfinish(
         logging.info("Shutting down and removing %s", name)
         emu.disconnect()
         emu.stop()
-        emu.delete()
+        if not session.config.getoption("avd_keep"):
+            emu.delete()
 
 
 def get_crash_reporter(pytestconfig):
